feat: Phase 1 int64 indexing - Shape, Slice, NDArray core types
Breaking change: Migrate from int32 to int64 for array indexing.

Core type changes:
- Shape: size, dimensions[], strides[], offset, bufferSize -> long
- Slice: Start, Stop, Step -> long
- SliceDef: Start, Step, Count -> long
- NDArray: shape, size, strides properties -> long/long[]

Helper methods:
- Shape.ComputeLongShape() for int[] -> long[] conversion
- Shape.Vector(long) overload

Related to #584

+++ src/NumSharp.Core/Backends/Unmanaged/UnmanagedStorage.Getters.cs
@@ -112,7 +112,7 @@ namespace NumSharp.Backends
         /// <seealso cref="GetView(Slice[])"/>
         /// <seealso cref="Alias()"/>
         [MethodImpl(OptimizeAndInline)]
-        public UnmanagedStorage GetData(params int[] indices)
+        public UnmanagedStorage GetData(params long[] indices)
         {
             var this_shape = Shape;
 
@@ -123,7 +123,7 @@ namespace NumSharp.Backends
                 var (shape, offset) = this_shape.GetSubshape(indices);
                 // For non-broadcasted contiguous subshapes, use size (the actual data extent).
                 // Only use BufferSize when the subshape itself is broadcasted.
-                int sliceSize = shape.IsBroadcasted
+                long sliceSize = shape.IsBroadcasted
                     ? (shape.BufferSize > 0 ? shape.BufferSize : shape.size)
                     : shape.size;
                 // Create shape with offset=0 since InternalArray.Slice already accounts for the offset
